+++ config/splat.us.dra.yaml
@@ -139,7 +139,7 @@ segments:
       - [0x41DD4, rodata] # EntityTeleport
       - [0x41E34, rodata] # EntitySubwpnCrashCross
       - [0x41E48, rodata]
-      - [0x41E5C, rodata] # func_80126ECC
+      - [0x41E5C, .rodata, 86ECC] # EntityHellfireHandler
       - [0x41E7C, rodata] # func_80128C2C
       - [0x41E90, rodata] # EntityStopWatch
       - [0x41EB0, .rodata, 8D3E8] # func_8012D3E8
@@ -187,6 +187,7 @@ segments:
       - [0x75F54, c, 75F54]
       - [0x7A4D0, c, 7A4D0]
       - [0x7E4BC, c, 7E4BC]
+      - [0x86ECC, c, 86ECC]
       - [0x8D3E8, c, 8D3E8]
       - [0x92F60, c, 92F60]
       - [0x93290, c, 93290]
