+++ config/symbols.hd.dra.txt
@@ -176,6 +176,10 @@ D_800AFFB8 = 0x800B007C;
 D_800B0130 = 0x800B01F4;
 D_800B01B8 = 0x800B027C;
 D_800B0608 = 0x800B06CC;
+D_800B06C8 = 0x800B078C;
+D_800B0728 = 0x800B07EC;
+D_800B0788 = 0x800B084C;
+D_800B0790 = 0x800B0854;
 D_800B0798 = 0x800B085C;
 D_800B07C8 = 0x800B088C;
 D_800B0830 = 0x800B08F4;
@@ -218,6 +222,7 @@ aXXVQ = 0x800DCC60;
 aAxearmor = 0x800DCC6C;
 aErrorStep = 0x800E1954;
 aAtariNuki = 0x800E19C0;
+aLightTimer02x = 0x800E1D64;
 func_800E346C = 0x800E3BEC;
 SetGameState = 0x800E4890;
 func_800E5358 = 0x800E58E4;
@@ -454,7 +459,7 @@ EntityHolyWaterBreakGlass = 0x80125AF4;
 EntityHolyWaterFlame = 0x80125F38;
 EntitySubwpnCrashCross = 0x8012643C;
 EntitySubwpnCrashCrossParticles = 0x801268D4;
-func_80126ECC = 0x80126B58;
+EntityHellfireHandler = 0x80126B58;
 func_801274DC = 0x80127168;
 func_8012768C = 0x80127318;
 func_80127840 = 0x801274CC;
