+++ config/symbols.us.dra.txt
@@ -268,6 +268,7 @@ EntityHolyWaterBreakGlass = 0x80125E68;
 EntityHolyWaterFlame = 0x801262AC;
 EntitySubwpnCrashCross = 0x801267B0;
 EntitySubwpnCrashCrossParticles = 0x80126C48;
+EntityHellfireHandler = 0x80126ECC;
 EntityExpandingCircle = 0x801279FC;
 EntitySubwpnReboundStone = 0x80127F40;
 EntitySubwpnThrownVibhuti = 0x80128714;
