+++ src/dra/7E4BC.c
@@ -951,687 +951,3 @@ void EntitySubwpnCrashCrossParticles(Entity* self) {
         }
     }
 }
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_80126ECC);
-
-void func_801274DC(Entity* entity) {
-    switch (entity->step) {
-    case 0:
-        if (entity->params == 0) {
-            PlaySfx(0x660);
-        }
-        entity->flags = FLAG_UNK_100000 | FLAG_UNK_08000000;
-        entity->animSet = ANIMSET_DRA(9);
-        entity->unk4C = &D_800B0798;
-        entity->zPriority = PLAYER.zPriority + 2;
-        entity->facingLeft = (PLAYER.facingLeft + 1) & 1;
-        SetSpeedX(D_800B0830[entity->params]);
-        entity->velocityY = D_800B083C[entity->params];
-        entity->ext.generic.unk7C.s = 0x14;
-        func_8011A328(entity, 2);
-        entity->hitboxWidth = 4;
-        entity->hitboxHeight = 4;
-        entity->step++;
-        break;
-
-    case 1:
-        if (entity->hitFlags == 0) {
-            entity->ext.generic.unk7C.s--;
-            if ((entity->ext.generic.unk7C.s) == 0) {
-                entity->step++;
-            }
-            entity->posX.val += entity->velocityX;
-            entity->posY.val += entity->velocityY;
-            break;
-        }
-        DestroyEntity(entity);
-        break;
-
-    case 2:
-        if (entity->hitFlags != 0) {
-            DestroyEntity(entity);
-            break;
-        }
-        entity->posX.val += entity->velocityX;
-        break;
-    }
-}
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_8012768C);
-
-void func_80127840(Entity* entity) {
-    switch (entity->step) {
-    case 0:
-        if (entity->params != 0) {
-            PlaySfx(0x683);
-        }
-
-        entity->flags = FLAG_UNK_100000 | FLAG_UNK_08000000;
-
-        if (entity->params != 0) {
-            entity->posY.i.hi = entity->posY.i.hi + 16;
-        } else {
-            entity->posY.i.hi = entity->posY.i.hi - 4;
-        }
-
-        entity->animSet = ANIMSET_DRA(9);
-        entity->rotZ = 0;
-        entity->unk4C = &D_800B07C8;
-        entity->drawFlags |= FLAG_DRAW_ROTZ;
-        entity->zPriority = PLAYER.zPriority + 2;
-        entity->facingLeft = (PLAYER.facingLeft + 1) & 1;
-        SetSpeedX(-0x10);
-        func_8011A328(entity, 2);
-        entity->hitboxWidth = 8;
-        entity->hitboxHeight = 8;
-        entity->step++;
-        break;
-
-    case 1:
-        if (entity->animFrameIdx >= 23) {
-            if (!(g_GameTimer & 3)) {
-                entity->rotZ += 0x400;
-            }
-            if (entity->velocityX < 0) {
-                entity->velocityX -= FIX(0.09375);
-            } else {
-                entity->velocityX += FIX(0.09375);
-            }
-            if (!(g_GameTimer & 1) && (rand() & 1)) {
-                CreateEntFactoryFromEntity(entity, FACTORY(0x100, 36), 0);
-            }
-            entity->posX.val += entity->velocityX;
-            entity->posY.val += entity->velocityY;
-        }
-        break;
-    }
-}
-
-// circle expands out of player
-void EntityExpandingCircle(Entity* entity) {
-    POLY_GT4* poly;
-    s32 primIndex;
-
-    if (PLAYER.facingLeft == 0) {
-        entity->posX.i.hi = PLAYER.posX.i.hi - 10;
-    } else {
-        entity->posX.i.hi = PLAYER.posX.i.hi + 10;
-    }
-    entity->posY.i.hi = PLAYER.posY.i.hi + 2;
-
-    switch (entity->step) {
-    case 0:
-        primIndex = AllocPrimitives(PRIM_GT4, 1);
-        entity->primIndex = primIndex;
-        if (primIndex != -1) {
-            entity->ext.generic.unk7C.s = 22;
-            entity->ext.generic.unk7E.modeU16 = 26;
-            poly = &g_PrimBuf[entity->primIndex];
-            poly->u2 = 64;
-            poly->u3 = 127;
-            poly->u1 = 127;
-            poly->v1 = 192;
-            poly->v0 = 192;
-            poly->v3 = 255;
-            poly->v2 = 255;
-            poly->r3 = 128;
-            poly->r2 = 128;
-            poly->r1 = 128;
-            poly->r0 = 128;
-            poly->g3 = 128;
-            poly->g2 = 128;
-            poly->g1 = 128;
-            poly->g0 = 128;
-            poly->u0 = 64;
-            poly->b3 = 64;
-            poly->b2 = 64;
-            poly->b1 = 64;
-            poly->b0 = 64;
-            poly->tpage = 0x1A;
-            poly->clut = 0x15F;
-            poly->pad2 = PLAYER.zPriority + 1;
-            poly->pad3 = 0x35;
-            entity->flags = FLAG_UNK_40000 | FLAG_UNK_04000000 | FLAG_HAS_PRIMS;
-            entity->step++;
-            break;
-        }
-        DestroyEntity(entity);
-        return;
-
-    case 1:
-        entity->ext.generic.unk7C.s += 2;
-        entity->ext.generic.unk7E.modeU16 += 2;
-        if (entity->ext.generic.unk7C.s >= 0x39) {
-            DestroyEntity(entity);
-            return;
-        }
-        break;
-
-    default:
-        break;
-    }
-
-    poly = &g_PrimBuf[entity->primIndex];
-    poly->x0 = entity->posX.i.hi - entity->ext.generic.unk7C.s;
-    poly->y0 = entity->posY.i.hi - entity->ext.generic.unk7E.modeU16;
-    poly->x1 = entity->posX.i.hi + entity->ext.generic.unk7C.s;
-    poly->y1 = entity->posY.i.hi - entity->ext.generic.unk7E.modeU16;
-    poly->x2 = entity->posX.i.hi - entity->ext.generic.unk7C.s;
-    poly->y2 = entity->posY.i.hi + entity->ext.generic.unk7E.modeU16;
-    poly->x3 = entity->posX.i.hi + entity->ext.generic.unk7C.s;
-    poly->y3 = entity->posY.i.hi + entity->ext.generic.unk7E.modeU16;
-
-    if (entity->ext.generic.unk7C.s >= 0x29) {
-        poly->r3 += 244;
-        poly->g3 += 244;
-        poly->b3 += 250;
-        poly->r0 = poly->r1 = poly->r2 = poly->r3;
-        poly->g0 = poly->g1 = poly->g2 = poly->g3;
-        poly->b0 = poly->b1 = poly->b2 = poly->b3;
-    }
-}
-
-void func_80127CC8(Entity* entity) {
-    POLY_GT4* poly;
-    s32 ret;
-
-    if (PLAYER.step != 34) {
-        DestroyEntity(entity);
-        return;
-    }
-
-    entity->posX.i.hi = PLAYER.posX.i.hi;
-
-    switch (entity->step) {
-    case 0:
-        ret = AllocPrimitives(PRIM_G4, 1);
-        entity->primIndex = ret;
-
-        if (ret == -1) {
-            DestroyEntity(entity);
-            g_Player.unk5C = 0xFFFF;
-            return;
-        }
-
-        entity->flags = FLAG_UNK_20000 | FLAG_UNK_40000 | FLAG_UNK_04000000 |
-                        FLAG_HAS_PRIMS;
-        poly = &g_PrimBuf[entity->primIndex];
-        poly->r3 = 192;
-        poly->r2 = 192;
-        poly->r1 = 192;
-        poly->r0 = 192;
-        poly->g3 = 64;
-        poly->g2 = 64;
-        poly->g1 = 64;
-        poly->g0 = 64;
-        poly->b3 = 64;
-        poly->b2 = 64;
-        poly->b1 = 64;
-        poly->b0 = 64;
-        poly->pad3 = 0x315;
-        entity->zPriority = 0x1C0;
-        poly->pad2 = 0x1C0;
-        entity->step++;
-        break;
-
-    case 1:
-        if (entity->ext.generic.unk7C.s++ >= 0xE) {
-            DestroyEntity(entity);
-            return;
-        }
-
-    default:
-        break;
-    }
-    poly = &g_PrimBuf[entity->primIndex];
-    poly->x0 = poly->x2 = entity->posX.i.hi - 3;
-    poly->y0 = 0;
-    poly->y1 = 0;
-    poly->x1 = poly->x3 = entity->posX.i.hi + 3;
-    poly->y3 = 240;
-    poly->y2 = 240;
-
-    if (g_GameTimer & 1) {
-        poly->pad3 = poly->pad3 | 8;
-    } else {
-        poly->pad3 = poly->pad3 & 0xFFF7;
-    }
-}
-
-void func_80127EAC(s16 arg0) {
-    g_CurrentEntity->ext.generic.unk7C.s =
-        (arg0 * 2) - g_CurrentEntity->ext.generic.unk7C.s;
-    if (g_CurrentEntity->ext.generic.unk80.modeS16.unk2 == 0) {
-        g_CurrentEntity->ext.generic.unk80.modeS16.unk0++;
-        g_CurrentEntity->ext.generic.unk80.modeS16.unk2++;
-    }
-}
-
-void func_80127EF0(s16 arg0) {
-    if (g_CurrentEntity->ext.generic.unk80.modeS16.unk2 == 0) {
-        g_CurrentEntity->ext.generic.unk7C.s =
-            (arg0 * 2) - g_CurrentEntity->ext.generic.unk7C.s;
-        g_CurrentEntity->ext.generic.unk80.modeS16.unk0++;
-        g_CurrentEntity->ext.generic.unk80.modeS16.unk2++;
-    }
-}
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnReboundStone);
-
-// ash thrown when using vibhuti subweapon
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnThrownVibhuti);
-
-s32 func_80128BBC(Unkstruct_80128BBC* arg0, u8 value) {
-    u8 ret = 0;
-    s32 i;
-    s32 j;
-    Unkstruct_80128BBC_Sub* temp = arg0->unk04;
-
-    for (i = 0; i < 4; i++, temp++) {
-        for (j = 0; j < 3; j++) {
-            temp->unk00[j] -= value;
-
-            if (temp->unk00[j] < 16) {
-                temp->unk00[j] = 16;
-            } else {
-                ret |= 1;
-            }
-        }
-    }
-    return ret;
-}
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_80128C2C);
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_801291C4);
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_80129864);
-
-// opens hole in backround and spirit comes out (ID 0x40)
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySummonSpirit);
-
-// expanding circle effect when activating stopwatch
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityStopWatchExpandingCircle);
-
-// stopwatch subweapon effect. stops enemies (Dra Entity 0x2A)
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityStopWatch);
-
-void func_8012B78C(Entity* entity) {
-    POLY_GT4* poly;
-    s32 ret;
-
-    switch (entity->step) {
-    case 0:
-        ret = AllocPrimitives(PRIM_GT4, 1);
-        entity->primIndex = ret;
-        if (entity->primIndex != -1) {
-            entity->flags = FLAG_UNK_20000 | FLAG_UNK_04000000 | FLAG_HAS_PRIMS;
-            poly = &g_PrimBuf[entity->primIndex];
-            poly->tpage = 0x1C;
-            poly->clut = 0x19D;
-            poly->u2 = 32;
-            poly->u0 = 32;
-            poly->u3 = 48;
-            poly->u1 = 48;
-            poly->v1 = 0;
-            poly->v0 = 0;
-            poly->v3 = 16;
-            poly->v2 = 16;
-            poly->x0 = poly->x2 = entity->posX.i.hi - 8;
-            poly->x1 = poly->x3 = entity->posX.i.hi + 8;
-            poly->y0 = poly->y1 = entity->posY.i.hi - 8;
-            poly->y2 = poly->y3 = entity->posY.i.hi + 8;
-            poly->pad2 = entity->zPriority;
-            poly->pad3 = 0x115;
-            entity->ext.generic.unk7E.modeU16 = 96;
-            entity->step++;
-        } else {
-            DestroyEntity(entity);
-            return;
-        }
-        break;
-
-    case 1:
-        if (++entity->ext.generic.unk7C.s > 5) {
-            entity->step++;
-        }
-        entity->ext.generic.unk7E.modeU16 -= 8;
-        break;
-
-    case 2:
-        DestroyEntity(entity);
-        return;
-
-    default:
-        break;
-    }
-    poly = &g_PrimBuf[entity->primIndex];
-    poly->r0 = poly->r1 = poly->r2 = poly->r3 = poly->g0 = poly->g1 = poly->g2 =
-        poly->g3 = poly->b0 = poly->b1 = poly->b2 = poly->b3 =
-            entity->ext.generic.unk7E.modeU8.unk0;
-}
-
-// book rotates around player
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntitySubwpnBible);
-
-// echo of bat effect
-INCLUDE_ASM("dra/nonmatchings/7E4BC", EntityBatEcho);
-
-INCLUDE_ASM("dra/nonmatchings/7E4BC", func_8012C600);
-
-bool WolfFormFinished(void) {
-    if (PLAYER.step_s == 0) {
-        return false;
-    }
-    if (PLAYER.step_s == 8) {
-        return false;
-    }
-    if (D_80097448[1] != 0 && !IsRelicActive(RELIC_HOLY_SYMBOL) ||
-        g_Player.padTapped & PAD_R2 ||
-        HandleTransformationMP(FORM_WOLF, REDUCE) < 0) {
-        SetPlayerStep(Player_Unk25);
-        func_8010DA48(0xCA);
-        D_800AFDA6[0] = 1;
-        PLAYER.palette = 0x810D;
-        g_Player.unk66 = 0;
-        g_Player.unk68 = 0;
-        CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x2400, 44), 0);
-        PLAYER.velocityY >>= 1;
-        return true;
-    }
-    return false;
-}
-
-void func_8012C97C(void) {
-    if (g_Entities[PLAYER_CHARACTER].step_s == 0) {
-        return;
-    }
-    if (g_Entities[PLAYER_CHARACTER].step_s >= 8 &&
-        g_Entities[PLAYER_CHARACTER].step_s < 10) {
-        return;
-    }
-    if (D_80097448[1] < 13) {
-        return;
-    }
-    if (!IsRelicActive(RELIC_HOLY_SYMBOL)) {
-        return;
-    }
-    if (!IsRelicActive(RELIC_SKILL_OF_WOLF)) {
-        return;
-    }
-    if (g_Player.pl_vram_flag & 1) {
-        return;
-    }
-    if (!(g_Player.padPressed & PAD_TRIANGLE)) {
-        return;
-    }
-    if (D_80138440 != 0) {
-        return;
-    }
-
-    PLAYER.step_s = 9;
-    D_800B0914 = 0;
-    func_8010DA48(0xEC);
-    PLAYER.velocityY = 0;
-}
-
-void func_8012CA64(void) {
-    u32 var_a0;
-
-    PLAYER.step_s = 1;
-    D_800B0914 = 0;
-
-    if (g_Player.pl_vram_flag & 0x20) {
-        var_a0 = 0xDF;
-    } else {
-        var_a0 = 0xDE;
-    }
-    func_8010DA48(var_a0);
-
-    PLAYER.velocityY = 0;
-    PLAYER.velocityX /= 2;
-
-    D_800B0918 = 0x200;
-    if (g_Player.pl_vram_flag & 0x40) {
-        D_800B0914 = 1;
-        func_8010DA48(0xE9U);
-    }
-}
-
-void func_8012CB0C(void) {
-    PLAYER.ext.generic.unkAC = 0xDE;
-    PLAYER.velocityY = 0;
-    D_800B0914 = 0;
-    PLAYER.animFrameIdx = 0;
-    PLAYER.animFrameDuration = 0;
-    PLAYER.step_s = 7;
-}
-
-void func_8012CB4C(void) {
-    PLAYER.step_s = 2;
-    if ((PLAYER.facingLeft != 0 && g_Player.padPressed & PAD_RIGHT) ||
-        (PLAYER.facingLeft == 0 && g_Player.padPressed & PAD_LEFT)) {
-        func_8010DA48(0xE1);
-        D_800B0914 = 0;
-        D_8013842C = 0;
-        return;
-    } else if (D_8013842C != 0) {
-        func_8010DA48(0xE2);
-        D_800B0914 = 2;
-        SetSpeedX(0x20000);
-        return;
-    } else {
-        func_8010DA48(0xE0);
-        D_800B0914 = 1;
-        D_8013842C = 0xC;
-    }
-}
-
-void func_8012CC30(s32 arg0) {
-    if (arg0 == 0) {
-        D_80138444 = 1;
-        if (g_ButtonCombo[COMBO_QCF].buttonsCorrect == COMBO_COMPLETE &&
-            IsRelicActive(RELIC_SKILL_OF_WOLF) &&
-            CastSpell(SPELL_WOLF_CHARGE)) {
-            func_8010E27C();
-            PLAYER.step_s = 2;
-            D_800B0914 = 4;
-            SetSpeedX(0x50000);
-            g_CurrentEntity->velocityY = 0;
-            func_8010DA48(0xEDU);
-            LearnSpell(4);
-        }
-    } else {
-        D_80138444 = 1;
-    }
-}
-
-void func_8012CCE4(void) {
-    PLAYER.velocityY = FIX(-3.5);
-    if ((PLAYER.step_s == 2) & (D_800B0914 == 2)) {
-        func_8010DA48(0xE7);
-        // Might be possible to rewrite this block to reduce duplication with
-        // some clever && and ||
-        if (PLAYER.facingLeft) {
-            if ((g_Player.pl_vram_flag & 0xF000) == 0xC000) {
-                PLAYER.velocityY = -(ABS(PLAYER.velocityX) + FIX(3.5));
-            }
-            if ((g_Player.pl_vram_flag & 0xF000) == 0x8000) {
-                PLAYER.velocityY = FIX(-0.5);
-            }
-        } else {
-            if ((g_Player.pl_vram_flag & 0xF000) == 0x8000) {
-                PLAYER.velocityY = -(ABS(PLAYER.velocityX) + FIX(3.5));
-            }
-            if ((g_Player.pl_vram_flag & 0xF000) == 0xC000) {
-                PLAYER.velocityY = FIX(-0.5);
-            }
-        }
-        D_800B0914 = 2;
-    } else if (
-        (g_Player.padPressed & (PAD_RIGHT | PAD_LEFT)) &&
-        ((PLAYER.step_s != 2) || (D_800B0914 != 0)) && (PLAYER.step_s != 9)) {
-        func_8010DA48(0xE7);
-        D_800B0914 = 1;
-        if (g_Player.padPressed & PAD_UP) {
-            PLAYER.velocityY = FIX(-4.875);
-        }
-    } else {
-        func_8010DA48(0xE6);
-        D_800B0914 = 0;
-        if (g_Player.padPressed & PAD_UP) {
-            PLAYER.velocityY = FIX(-4.875);
-        }
-    }
-    PLAYER.step_s = 4;
-    D_80138430 += 0x80;
-}
-
-void func_8012CED4(void) {
-    if (PLAYER.step_s == 2 && D_800B0914 == PLAYER.step_s) {
-        func_8010DA48(0xE7);
-        D_800B0914 = 1;
-    } else {
-        func_8010DA48(0xE8);
-        SetSpeedX(0x10000);
-        D_800B0914 = 0;
-        if (D_80138438 & 0x40) {
-            PLAYER.animFrameIdx = 4;
-            PLAYER.velocityX = 0;
-            PLAYER.animFrameDuration = 1;
-        }
-    }
-    PLAYER.step_s = 5;
-    g_Player.D_80072F0A = 8;
-    PLAYER.velocityY = 0;
-    D_80138430 -= 0x100;
-}
-
-void func_8012CFA8(void) {
-    func_8010DA48(0xEA);
-    PLAYER.step_s = 6;
-    D_800B0914 = 0;
-    PLAYER.velocityX = 0;
-    g_Player.D_80072F0A = 8;
-}
-
-void func_8012CFF0(void) {
-    PLAYER.step_s = 3;
-    func_8010DA48(0xE3);
-    D_800B0914 = 0;
-}
-
-void func_8012D024(void) {
-    s32 separation;
-    DecelerateX(0x2000);
-    if (g_Player.padTapped & PAD_CROSS) {
-        func_8012CCE4();
-        return;
-    }
-    if (!(g_Player.pl_vram_flag & 1)) {
-        func_8012CED4();
-        return;
-    }
-    if (g_Player.padTapped & PAD_SQUARE) {
-        func_8012CC30(0);
-        return;
-    }
-    if (g_Player.padPressed & (PAD_RIGHT | PAD_LEFT)) {
-        func_8012CB4C();
-        return;
-    }
-    if (g_Player.padPressed & PAD_DOWN) {
-        func_8012CFF0();
-        return;
-    }
-    if (D_800B0914 != 0) {
-        return;
-    }
-    separation = (PLAYER.posY.i.hi - g_Entities[17].posY.i.hi);
-    if (ABS(separation) < 4 && --D_800B0918 == 0) {
-        D_800B0914 = 1;
-        func_8010DA48(0xE9);
-        return;
-    }
-    if (g_Player.pl_vram_flag & 0x40) {
-        D_800B0914 = 1;
-        func_8010DA48(0xE9);
-    }
-}
-
-void func_8012D178(void) {
-    s32 var_v0;
-
-    if (g_Player.padTapped & PAD_CROSS) {
-        func_8012CCE4();
-    } else if (!(g_Player.pl_vram_flag & 1)) {
-        func_8012CFA8();
-    } else {
-#if defined(VERSION_US)
-        if (PLAYER.facingLeft != 0) {
-            var_v0 = g_Player.padPressed & PAD_LEFT;
-        } else {
-            var_v0 = g_Player.padPressed & PAD_RIGHT;
-        }
-#elif defined(VERSION_HD)
-        var_v0 = g_Player.padPressed & (PAD_LEFT | PAD_RIGHT);
-#endif
-        if (var_v0 != 0) {
-            func_8012CB4C();
-        } else if (g_Player.unk04 & 0x40) {
-            func_8012CA64();
-        } else if (g_GameTimer % 6 == 0) {
-            CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x100, 69), 0);
-        }
-    }
-}
-
-void func_8012D28C(bool exitEarly) {
-    bool bitNotFound;
-    s32 i;
-
-    func_80102CD8(2);
-    PlaySfx(0x644);
-    PLAYER.velocityX = 0;
-    g_Player.D_80072EFC = 0x20;
-    g_Player.D_80072EF4 = 0;
-    // Odd logic, if we exit early, we force an R2-tap. Strange!
-    if (exitEarly) {
-        g_Player.padTapped = PAD_R2;
-        WolfFormFinished();
-        return;
-    }
-    // Start a routine where we look through this array for a value.
-    bitNotFound = 0;
-    for (i = 3; i < 7; i++) {
-        if ((g_Player.D_80072CF0[i][0] & 2)) {
-            break;
-        }
-    }
-    // If we made it through that loop without finding one, skip i=7,8,9
-    // and keep searching.
-    if (i == 7) {
-        for (i = 10; i < 14; i++) {
-            if ((g_Player.D_80072CF0[i][0] & 2)) {
-                break;
-            }
-        }
-    }
-    // If we even made it through that one, then conclude the bit was not found.
-    if (i == 14) {
-        bitNotFound++;
-    }
-
-    SetSpeedX(FIX(1));
-    CreateEntFactoryFromEntity(g_CurrentEntity, FACTORY(0x900, 4), 0);
-    D_800B0914 = 0;
-    // Finally make use of that bit to control if X is positive or negative.
-    if (bitNotFound) {
-        func_8012CED4();
-        SetSpeedX(FIX(1));
-    } else {
-        func_8012CFA8();
-        SetSpeedX(FIX(-1));
-    }
-    PLAYER.velocityY = FIX(-3.5);
-}
