+++ src/dra/dra.h
@@ -493,6 +493,10 @@ extern s16 D_800AFFB8[];
 extern s32 D_800B0130[];
 extern s32 D_800B01B8[];
 extern u8 D_800B0608[];
+extern u32 D_800B06C8;
+extern u32 D_800B0728;
+extern RECT D_800B0788;
+extern RECT D_800B0790;
 extern s32 D_800B0798[];
 extern s32 D_800B07C8;
 extern s32 D_800B0830[];
@@ -535,6 +539,7 @@ extern Unkstruct_800BF554 D_800BF554[];
 
 extern char* aErrorStep;
 extern char* aAtariNuki;
+extern char* aLightTimer02x;
 extern s32 g_DebugFreeze;
 extern s32 g_DebugHitboxViewMode;
 extern u32 D_801362B4;
@@ -1013,7 +1018,7 @@ void func_80125E68(Entity* entity);
 void func_801262AC(Entity* entity);
 void func_801267B0(Entity* entity);
 void func_80126C48(Entity* entity);
-void func_80126ECC(Entity* entity);
+void EntityHellfireHandler(Entity* entity);
 void func_801274DC(Entity* entity);
 void func_8012768C(Entity* entity);
 void func_80127840(Entity* entity);
